% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPitchTier.R
\name{extractPitchTier}
\alias{extractPitchTier}
\title{extractPitchTier}
\usage{
extractPitchTier(x = "A01", res = 0.1, fmin = 50, fmax = 250,
  stylize = 1, wd = getwd())
}
\arguments{
\item{x}{name of wav file}

\item{res}{resolution (Praat)}

\item{fmin}{minimum frequency (Praat)}

\item{fmax}{maximum frequency (Praat)}

\item{stylize}{stylize (Praat)}

\item{wd}{working directory for Praat to use}
}
\value{
Runs script in Praat and prints stdout to console.
}
\description{
Extracts Praat PitchTier from wav object.
}
