% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractWord.R
\name{extractWord}
\alias{extractWord}
\title{extractWord}
\usage{
extractWord(
  x,
  word,
  tier = "Word",
  ignore_case = TRUE,
  instance = "random",
  wd = getwd()
)
}
\arguments{
\item{x}{path to a TextGrid}

\item{word}{word to search for}

\item{tier}{name of word tier in TextGrid}

\item{ignore_case}{default is `TRUE`}

\item{instance}{instance of word in TextGrid to extract. Default extracts a random instance. Can also be numeric (row number)}

\item{wd}{working directory for Praat to use. Accepts relative paths.}
}
\value{
Extracts section of wav file corresponding to word and saves in format name_wordi.wav where name is the original name, word is the word and x is the numeric instance.
}
\description{
Extract from a wav file with reference to a TextGrid.
}
\seealso{
density
}
