% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findPeak.R
\name{findPeak}
\alias{findPeak}
\title{findPeak}
\usage{
findPeak(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{further arguements to be passed to `density`}
}
\value{
Returns the value of `x` that corresponds to the peak of the density curve.
}
\description{
Call Praat via system2()
}
\seealso{
density
}
