% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findWord.R
\name{findWord}
\alias{findWord}
\title{findWord}
\usage{
findWord(x, word = "speech", tier = "Word", ignore_case = TRUE)
}
\arguments{
\item{x}{path to a TextGrid}

\item{word}{word to search for}

\item{tier}{name of word tier in TextGrid}

\item{ignore_case}{default is `TRUE`}
}
\value{
Returns a tibble with onset (t1) and offset (t2) of each occurance of the word in the TextGrid
}
\description{
Find a word in a TextGrid
}
\seealso{
extractWord
}
