% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatF0.R
\name{flatF0}
\alias{flatF0}
\title{flatF0}
\usage{
flatF0(wav, .f = findPeak, ...)
}
\arguments{
\item{wav}{path to a wav file or a tuneR WAVE object}

\item{.f}{function to use to determine pitch. Default is findPeak which finds the mode of the existing pitch contour.}

\item{...}{Additional arguments passed to extractPitchTier}
}
\value{
Returns a tuneR WAVE object of the input with a flat F0 contour
}
\description{
Flatten fundamental frequency contour using 'Praat'
}
\seealso{
extractPitchTier
}
