% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length.R
\name{length,Wave-method}
\alias{length,Wave-method}
\title{S4 generic for length}
\usage{
\S4method{length}{Wave}(x)
}
\arguments{
\item{x}{a 'tuneR' WAVE object}
}
\value{
The length of the left channel of the WAVE object
}
\description{
S4 generic for length.
}
\seealso{
length
}
