% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mm1}
\alias{mm1}
\title{mm1}
\format{
A tuneR "Wave" object:
}
\usage{
mm1
}
\description{
Example speech from Tilsen & Johnson (2008)
}
\references{
Tilsen, S., & Johnson, K. (2008). Low-frequency Fourier analysis of speech rhythm. The Journal of the Acoustical Society of America, 124(2), EL34–EL39. doi:10.1121/1.2947626
}
\keyword{datasets}
