% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatScript.R
\name{praatScript}
\alias{praatScript}
\title{praatScript}
\usage{
praatScript(args, script = "reTimeWin.praat", wd = getwd(),
  praat = NULL)
}
\arguments{
\item{args}{arguements to pass to Praat script ("--run" not required)}

\item{script}{name of script if using a script from this package, or path to script for other scripts}

\item{wd}{working directory for Praat to use}

\item{praat}{path to Praat. If null will search for Praat in C:/Program Files (for Windows) or attempt to use "praat" for Unix based systems.}
}
\value{
Runs script in Praat and prints stdout to console.
}
\description{
Executes a Praat script using the R system function.
}
