% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/praatSys.R
\name{praatSys}
\alias{praatSys}
\title{praatSys}
\usage{
praatSys(args = "--version", praat = NULL, ...)
}
\arguments{
\item{args}{arguements to pass to 'Praat'}

\item{praat}{path to 'Praat'. If null will search for 'Praat' in C:/Program Files (for Windows) or attempt to use "praat" for Unix based systems.}

\item{...}{arguements to pass to internal get_praat_path() function. This can be used to change the folder to look for R in for Windows (default is appDir = "C:/Program Files")}
}
\value{
Prints stdout to console
}
\description{
Call 'Praat' via system2()
}
\seealso{
system2
}
