% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_tg.R
\name{read_tg}
\alias{read_tg}
\title{read_tg}
\usage{
read_tg(file)
}
\arguments{
\item{file}{path to TextGrid file}
}
\value{
Returns a nested tibble with `name`, `type` and `data`. `data` has the variables `t1`, `t2` and `label`
}
\description{
Reads a Praat TextGrid as a nested tibble
}
