% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRetime.R
\name{runRetime}
\alias{runRetime}
\title{runRetime}
\usage{
runRetime(audio, grid, outfile, wd = getwd())
}
\arguments{
\item{audio}{relative path to input audio file}

\item{grid}{relative path to input TextGrid}

\item{outfile}{relative path to output audio file}

\item{wd}{path to working directory}
}
\value{
A wav file with the timing of the second tier of the TextGrid will be saved to the outfile location.
}
\description{
runRetime
}
