% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tempogram.R
\name{tempogram}
\alias{tempogram}
\title{tempogram}
\usage{
tempogram(x, sr = 250, window = 200, hop = 100)
}
\arguments{
\item{x}{signal as vector}

\item{sr}{sample rate}

\item{window}{length of acf window}

\item{hop}{distance between windows}
}
\value{
a list including:
   acf: tempogram in long format tibble
   tempo: peak tempos for each time bin (f0 of tempogram)
   acf.m: tempogram in 2D matrix
}
\description{
Produce a tempogram of a time series. No longer exported due to the rray requirement
}
