% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsola.R
\name{wsola}
\alias{wsola}
\title{wsola}
\usage{
wsola(x, s, win = "hann", winLen = 1024, synHop = 512, tol = 512)
}
\arguments{
\item{x}{an audio signal}

\item{s}{a scaling factor or a list of two vector with anchor points}

\item{win}{window function. Default is 'hann' for hanning window. Can also be a custom window supplied as a vector}

\item{winLen}{window length}

\item{synHop}{synthesis window hop size}

\item{tol}{tolerance for overlap delta}
}
\value{
retimed audio signal as vector
}
\description{
Waveform Similarity Overlap-add. Translated from TSM Toolbox.
}
\references{
Driedger, J., Müller, M., Fourier, J., & Shannon, C. (2014). TSM Toolbox: MATLAB Implementations of Time-Scale Modification Algorithms. 8.
}
\seealso{
fft_spectrum, get_serials_anchors
}
