% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPitchTier.R
\name{extractPitchTier}
\alias{extractPitchTier}
\title{extractPitchTier}
\usage{
extractPitchTier(wav, res = 0.1, fmin = 50, fmax = 250, output = "PitchTier")
}
\arguments{
\item{wav}{path to a wav file or a tuneR WAVE object}

\item{res}{resolution of PitchTier}

\item{fmin}{minimum frequency of PitchTier}

\item{fmax}{maximum frequency of PitchTier}

\item{output}{can be "PitchTier" or "file"}
}
\value{
Returns a PitchTier object or the temporary path to the generated PitchTier file
}
\description{
Extracts 'Praat' PitchTier from wav object.
}
