% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatF0.R
\name{flatF0}
\alias{flatF0}
\title{flatF0}
\usage{
flatF0(filename, .f = findPeak, wd = getwd())
}
\arguments{
\item{filename}{name of wav file to flatten}

\item{.f}{function to use to determine pitch}

\item{wd}{working directory for praat to use}
}
\value{
Creates wav file with flat F0 contour.
}
\description{
Flatten pitch
}
\seealso{
extractPitchTier
}
