% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retimer-package.R
\docType{package}
\name{retimer-package}
\alias{retimer}
\alias{retimer-package}
\title{retimer: Retime and Repitch Speech}
\description{
Tools used for retiming and analysis of speech with R and Praat.
}
\author{
\strong{Maintainer}: Alistair Beith \email{alistair.beith@gmail.com}

}
\keyword{internal}
