% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrogram.R
\name{spectrogram}
\alias{spectrogram}
\title{spectrogram}
\usage{
spectrogram(
  x,
  fs = NULL,
  method = NULL,
  output = "tibble",
  wintime = 25,
  steptime = 10
)
}
\arguments{
\item{x}{a signal, `tuneR` WAVE object, or the path to an .wav or .mp3 file.}

\item{fs}{sample rate if supplying the signal as a vector}

\item{method}{spectrogram implementation to use. Available options are 'phonTools', 'tuneR', 'gsignal', and 'seewave'. Default is to select the first of these methods that is available.}

\item{output}{format of output}

\item{wintime}{length of analysis window in ms}

\item{steptime}{interval between steps in ms}
}
\value{
Returns a spectrogram in the desired format
}
\description{
Universal spectrogram function.
}
