% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_tg.R
\name{write_tg}
\alias{write_tg}
\title{write_tg}
\usage{
write_tg(x, file)
}
\arguments{
\item{x}{Nested tibble. Must contain the columns `name`, `type` and `data`. `data` must have the columns `t1`, `t2` and `label`}

\item{file}{File name to save TextGrid as}
}
\value{
Returns path of saved TextGrid file
}
\description{
Writes a nested tibble to a "Praat" TextGrid file
}
